updated
HTML: For the structure and layout of the game board.
CSS: For styling the game elements and providing a visually appealing interface.
JavaScript: For implementing the game logic, dice rolling, player movement, and handling game events.
Git: For version control and collaboration with others on the project.
and coded with VS code
This project is a fun way to practice JavaScript programming skills and to enjoy a classic board game with friends or family online. Feel free to fork and contribute to this project to add more features or improve existing ones. Let's have some nostalgic fun playing the Snake and Ladder game together! 🐍🪜